I added to the main_function_convnets script the function linear_and_forward_activation which makes the linear activation and applies the activation function in a fully connected layer. I made this function instead of keeping with the function of the main_functions_classic NN script because in the new function I get both values of Z and A instead of just the value of A.
On the other hand, I kept with the function forward_propagation_convnet. I have made the several cases of the type of layer each network could have. I also raised and else statement in which if the name of the current layer is not conv, flatten or fully_connected I raise an Exception error.